+++ backend/index.ts
@@ -5,6 +5,7 @@ import config from './config';
 import authRoutes from './routes/auth'; // Import the auth routes
 import salaryRoutes from './routes/salary';
 import employeeRoutes from './routes/employee'
+import advanceRoutes from './routes/advance';
 
 
 const app = express();
@@ -18,7 +19,8 @@ app.use(express.json());
 
  app.use('/auth', authRoutes);
 app.use('/salary', salaryRoutes);
-app.use('/employees',employeeRoutes);
+app.use('/employee',employeeRoutes);
+app.use('/advance',advanceRoutes);
 
 
 app.listen(config.port, () => {
