+++ frontend/package.json
@@ -13,6 +13,7 @@
     "@emotion/react": "^11.11.3",
     "@emotion/styled": "^11.11.0",
     "@mui/material": "^5.15.11",
+    "@mui/x-date-pickers": "^6.19.5",
     "axios": "^1.6.7",
     "dotenv": "^16.4.5",
     "react": "^18.2.0",
