+++ frontend/src/Router.tsx
@@ -3,15 +3,15 @@ import { Page } from './components/Page';
 import { Dashboard } from './pages/Dashboard';
 import LoginForm from './pages/Auth/LoginForm';
 import SignUpForm from './pages/Auth/SignUpForm';
-import AddEmployee from './pages/AddEmployee.tsx';
-import SalaryIssued from './pages/SalaryIssued.tsx';
-import IssuedAdvance from './pages/IssuedAdvance';
-import ListOfEmployee from './pages/ListOfEmployee';
-import AddExpense from './pages/AddExpense';
-import RequestSalary from './pages/RequestSalary';
-import ListOfExpense from './pages/ListOfExpense';
-import Wallet from './pages/Wallet';
-import HistoryComponent from './pages/History';
+import AddEmployee from './pages/Admin/AddEmployee.tsx';
+import SalaryIssued from './pages/Admin/SalaryIssued.tsx';
+import IssuedAdvance from './pages/Admin/IssuedAdvance.tsx';
+import ListOfEmployee from './pages/Admin/ListOfEmployee.tsx';
+import AddExpense from './pages/employee/AddExpense.tsx';
+import RequestSalary from './pages/employee/RequestSalary.tsx';
+import ListOfExpense from './pages/employee/ListOfExpense.tsx';
+import Wallet from './pages/employee/Wallet.tsx';
+import HistoryComponent from './pages/History.tsx';
 import Profile from './pages/Profile';
 
 function Router() {
